+++ src/components/single-experiment/accessibility.tsx
@@ -6,7 +6,7 @@ import { absoluteURL, getErrorMessage } from "@/lib/utils";
 import { Share2 } from "lucide-react";
 import { toast } from "sonner";
 
-export default function ExperimentInfo({experiment, slug}: SingleExperimentComponentProps & {
+export default function ExperimentAccessibility({experiment, slug}: SingleExperimentComponentProps & {
      slug: string
 }){
      const shareExperiment = async() => {
@@ -23,7 +23,7 @@ export default function ExperimentInfo({experiment, slug}: SingleExperimentCompo
           }
      }
      return (
-          <div className="sticky top-17.25 w-full not-prose bg-background border-b p-2 flex justify-between items-center">
+          <div className="sticky top-17.5 w-full not-prose bg-background border-b p-2 flex justify-between items-center z-30">
                <Breadcrumb className="w-3/5">
                     <BreadcrumbList className="text-xs sm:text-sm">
                          <BreadcrumbItem>